Rewrite using java implementation for PollForResponse / PollForErrorResponse / Poll

+++ aeron/aeron.go
@@ -29,7 +29,7 @@ import (
 )
 
 // NullValue is used to represent a null value for when some value is not yet set.
-const NullValue = int64(-1)
+const NullValue = -1
 
 // NewPublicationHandler is the handler type for new publication notification from the media driver
 type NewPublicationHandler func(string, int32, int32, int64)
